Add sidebar section (#247)
This addresses #246 and introduces a sidebar with plugin content outlines. To implement the sidebar, several considerations were taken into account:
* Keep the sidebar isolated from the header and footer
* Disable the sidebar for small screens (mobile phones)
* Make it in a fixed position, and allow for scrolling.

We used Materials UI `Drawer` component, it handles most of the above with less code.

Things are done in `App.jsx` and modulized to the have components in `components` folder.

The project folder structure
```
- public
- src
 -- assets
 -- Components
      ---MainIndex.jsx
      ---Details.jsx
      ---Sidebar.jsx
  --App.jsx
  --main.jsx
-index.html
```
