+++ aiida-registry-app/src/App.css
@@ -1,9 +1,20 @@
 #root {
-
   text-align: left;
 }
 html {
   height: 100%;
+}
+
+header {
+  background-color: black;
+  transition: top 0.5s;
+  display: block;
+  position: fixed;
+  top: 0;
+  z-index: 1;
+  margin: 0;
+  width: 100%;
+  padding: 5px 0px 14px 0px;
 }
 
 body {
@@ -23,6 +34,8 @@ main {
   padding-right: 0px;
   padding-bottom: 0px;
   margin-bottom: 0px;
+  max-width: 1280px;
+  margin: 0 auto;
 }
 
 p {
@@ -51,6 +64,19 @@ h1 {
   background-color: #d5cccc;
 }
 
+#detailsContainer {
+  display: flex;
+  margin-left: 50px;
+}
+
+#sidebar .MuiDrawer-paper {
+  width: 340px;
+  max-height:calc(100vh);
+  background-color:lightgray;
+  border: 5px solid gray;
+  transition: 0.5s;
+  margin-top: 155px;
+}
 .keyword {
   display: inline;
   padding: .2em .6em .3em;
@@ -70,7 +96,6 @@ h1 {
 #app-body {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
 }
 
 .plugin-header, .plugin-header a {
@@ -88,8 +113,9 @@ h2 {
   color: #333;
 }
 
-.details {
+#details {
   margin-top: 12px;
+  width: calc(100% - 370px);
 }
 
 .currentstate {
@@ -179,8 +205,8 @@ td {
 
 .footer {
   display: table;
-  width: calc(100% - 64px);
   /* Distance with other content */
+  width: calc(100% - 64px);
   margin-top: 16px;
   padding-top: 0.5rem;
   padding-bottom: 0.5rem;
